fbos (rendering still not working) + some fixes

+++ src/io/window.h
@@ -8,13 +8,13 @@ typedef struct {
     GLFWwindow* glfw_window;
 
     char* title;
-    u16 width;
-    u16 height;
+    u32 width;
+    u32 height;
 
     bool cursor_hidden;
 } window_s;
 
-void create_window(u16 width, u16 height, char* title);
+void create_window(u32 width, u32 height, char* title);
 void destroy_window();
 
 void window_set_cursor_visibility(window_s* window, bool visibility);
